pyproject.toml
[tool.poetry]
name = "new-bot"
version = "0.1.0"
description = ""
authors = ["rustam95ur <rustam95ur>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
